+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 ## Changelog
 
 - [command_schedule] improvements to this module.
+  - feature: rename the fields in the config for clarity. (The old config schema will be migrated automatically)
   - feature: optimize the `default config` for this module, provide an intuitive default example, make it easier to understand.
+  - docs: document how this module works.